+++ py_css_styleguide/nomenclature.py
@@ -1,5 +1,26 @@
 """
-Manifest nomenclature
+Nomenclature
+============
+
+* A rule name starts with a prefix from ``RULE_BASE_PREFIX``;
+* Rule prefix is followed from its type that can be either:
+
+  * ``RULE_META``;
+  * ``reference``;
+
+* Then rule name ends with its component name;
+* Rule name parts are separated with ``-``.
+
+There is a limited set of allowed characters:
+
+* For rule component name they are registred in ``RULE_ALLOWED_CHARS``;
+* For rule property name they are registred in ``PROPERTY_ALLOWED_CHARS``;
+
+There is some reserved name:
+
+* For rule component name they are registred in ``RESERVED_RULE_NAMES``;
+* For rule property name they are registred in ``RESERVED_PROPERTY_NAMES``;
+
 """
 from string import ascii_letters, digits