+++ tests/conftest.py
@@ -30,3 +30,12 @@ class FixturesSettingsTestMixin(object):
 def fixtures_settings():
     """Initialize and return settings (mostly paths) for fixtures (scope at module level)"""
     return FixturesSettingsTestMixin()
+
+
+@pytest.fixture(scope='session')
+def temp_builds_dir(tmpdir_factory):
+    """
+    Prepare a temporary build directory
+    """
+    fn = tmpdir_factory.mktemp('py_css_styleguide-tests')
+    return fn